clean up some ImageCache stuff

+++ src/SmallSmallRSS/ImageCache.php
@@ -3,14 +3,24 @@ namespace SmallSmallRSS;
 
 class ImageCache
 {
+    public static function hashToFilename($hash)
+    {
+        return \SmallSmallRSS\Config::get('CACHE_DIR') . '/images/' . $hash . '.png';
+    }
+
+    public static function hashExists($hash)
+    {
+        return file_exists(self::hashToFilename($hash));
+    }
+
     public static function processImages($html, $site_url, $download_if_missing = true)
     {
         $cache_dir = \SmallSmallRSS\Config::get('CACHE_DIR') . "/images";
         libxml_use_internal_errors(true);
         $charset_hack = '<head><meta http-equiv="Content-Type" content="text/html; charset=utf-8"/></head>';
-        $doc = new DOMDocument();
+        $doc = new \DOMDocument();
         $doc->loadHTML($charset_hack . $html);
-        $xpath = new DOMXPath($doc);
+        $xpath = new \DOMXPath($doc);
         $entries = $xpath->query('(//img[@src])');
         foreach ($entries as $entry) {
             self::processEntry($entry, $site_url);
@@ -29,24 +39,19 @@ class ImageCache
             return;
         }
         $src = \SmallSmallRSS\Utils::rewriteRelativeUrl($site_url, $src);
-        $cached = self::isCached($src, $download_if_missing);
+        $cached = self::isCached($src, $download_if_missing, $site_url);
         if ($cached) {
             $src = \SmallSmallRSS\Config::get('SELF_URL_PATH') . '/image.php?hash=' . sha1($src);
         }
         $entry->setAttribute('src', $src);
     }
 
-    public static function hashExists()
+    public static function isCached($src, $download_if_missing = true, $site_url = false)
     {
-        $filename = \SmallSmallRSS\Config::get('CACHE_DIR') . '/images/' . $hash . '.png';
-        return file_exists($filename);
-    }
-
-    public static function isCached($src, $download_if_missing = true)
-    {
-        $local_filename = \SmallSmallRSS\Config::get('CACHE_DIR') . "/images/" . sha1($src) . ".png";
-        if (!file_exists($local_filename)) {
-            $file_content = \SmallSmallRSS\Fetcher::fetch($src);
+        $local_filename = self::hashToFilename(sha1($src));
+        if (!file_exists($local_filename) && $download_if_missing) {
+            $fetcher = new \SmallSmallRSS\Fetcher($site_url);
+            $file_content = $fetcher->fetch($src);
             if ($file_content && strlen($file_content) > 1024) {
                 file_put_contents($local_filename, $file_content);
                 return true;
